ws sample fixed up
