+++ samples/dotnet/websockets/Agents/Agent.Core/AIHelpers.cs
@@ -1,4 +1,4 @@
-﻿namespace wsAgent.Core;
+﻿namespace Agent.Core;
 using System;
 using System.Collections.Immutable;
 using System.Net.WebSockets;
