+++ samples/dotnet/websockets/Agents/Agent.Core/Agent.Core.csproj
@@ -2,12 +2,12 @@
 
 	<ItemGroup>
 		<PackageReference Include="Azure.Identity" />
-		<PackageReference Include="Microsoft.AspNetCore.SignalR.Client" />
 		<PackageReference Include="Microsoft.SemanticKernel" />
 		<PackageReference Include="System.Linq.Async" />
 	</ItemGroup>
 
 	<ItemGroup>
+		<ProjectReference Include="..\..\..\..\..\src\Client\Yaap.Client.Abstractions.csproj" />
 		<ProjectReference Include="..\..\ApiClient\ApiClient.csproj" />
 		<ProjectReference Include="..\..\Assistants\Assistants.csproj" />
 		<ProjectReference Include="..\..\Models\Models.csproj" />
