+++ samples/dotnet/websockets/Agents/Agent.Core/DefaultAgentFunctionInvocationFilter.cs
@@ -1,18 +1,19 @@
 ﻿namespace Agent.Core;
 
+using System.Net.WebSockets;
+using System.Text;
 using System.Text.Json;
 
 using Common;
 
-using Microsoft.AspNetCore.SignalR.Client;
 using Microsoft.Extensions.Logging;
 
-public class DefaultFunctionInvocationFilter(ILogger log, HubConnection signalr) : FunctionInvocationFilter(
+internal class DefaultFunctionInvocationFilter(ILogger log, ClientWebSocket signalr) : FunctionInvocationFilter(
         onStart: async context =>
         {
             var msg = $"Running {context.Function.Name} ({context.Function.Description}) ...";
             log.LogTrace(msg);
-            await signalr.SendAsync(Constants.SignalR.Functions.PostStatus, msg);
+            signalr.SendAsync(new ArraySegment<byte>(Encoding.UTF8.GetBytes($@"{{ ""message"": ""{msg}"" }}")), WebSocketMessageType.Text, true, CancellationToken.None);
         },
         onEnd: async context =>
         {
@@ -27,20 +28,20 @@ public class DefaultFunctionInvocationFilter(ILogger log, HubConnection signalr)
                 serializedValue = context.Result.ToString();
             }
 
-            log.LogTrace("{msg} Result: {result}", msg, serializedValue);
-            await signalr.SendAsync(Constants.SignalR.Functions.PostStatus, msg);
+            log.MsgResultResult(msg, serializedValue);
+            signalr.SendAsync(new ArraySegment<byte>(Encoding.UTF8.GetBytes($@"{{ ""message"": ""{msg}"" }}")), WebSocketMessageType.Text, true, CancellationToken.None);
         },
         onAutoStart: async context =>
         {
             var msg = $"Running AUTO {context.Function.Name} ({context.Function.Description}) ...";
             log.LogTrace(msg);
-            await signalr.SendAsync(Constants.SignalR.Functions.PostStatus, msg);
+            signalr.SendAsync(new ArraySegment<byte>(Encoding.UTF8.GetBytes($@"{{ ""message"": ""{msg}"" }}")), WebSocketMessageType.Text, true, CancellationToken.None);
         },
         onAutoEnd: async context =>
         {
             var msg = $"AUTO {context.Function.Name} completed.";
-            log.LogTrace("{msg} Result: {result}", msg, context.Result.ToString());
-            await signalr.SendAsync(Constants.SignalR.Functions.PostStatus, msg);
+            log.MsgResultResult(msg, context.Result.ToString());
+            signalr.SendAsync(new ArraySegment<byte>(Encoding.UTF8.GetBytes($@"{{ ""message"": ""{msg}"" }}")), WebSocketMessageType.Text, true, CancellationToken.None);
         }
 )
 { }
