+++ samples/dotnet/websockets/Agents/Orchestrator/Dockerfile
@@ -3,6 +3,9 @@ RUN apt-get update \
     && apt-get upgrade -y
 USER app
 
+ARG AZURE_OPENAI_ENDPOINT
+ENV AzureOpenAIEndpoint=$AZURE_OPENAI_ENDPOINT
+
 ARG AZURE_OPENAI_KEY
 ENV AzureOpenAIKey=$AZURE_OPENAI_KEY
 
@@ -10,4 +13,4 @@ WORKDIR /app
 
 COPY . .
 
-ENTRYPOINT ["dotnet", "Orchestrator_WS.dll"]
+ENTRYPOINT ["dotnet", "Orchestrator.dll"]
