+++ samples/dotnet/websockets/Agents/Orchestrator/appsettings.Development.json
@@ -1,7 +1,7 @@
 {
   "Logging": {
     "LogLevel": {
-      //"Orchestrator_WS": "Trace",
+      "Orchestrator": "Trace",
       "Yaap": "Trace",
       "Default": "Information",
       "Microsoft.AspNetCore": "Warning",
