--- samples/dotnet/websockets/Agents/SignalR/Districts_SignalR/DistrictApi_Extras.cs
@@ -1,14 +1,0 @@
-﻿namespace TBAAPI.V3Client.Api;
-
-using System.Text.Json;
-
-using Microsoft.Extensions.Logging;
-
-using TBAAPI.V3Client.Client;
-
-public partial class DistrictApi
-{
-    //private ILogger? Log { get; }
-
-    //public DistrictApi(Configuration config, ILogger logger) : this(config) => this.Log = logger;
-}
