+++ samples/dotnet/websockets/Agents/TeamsExpert/Dockerfile
@@ -3,6 +3,9 @@ RUN apt-get update \
     && apt-get upgrade -y
 USER app
 
+ARG AZURE_OPENAI_ENDPOINT
+ENV AzureOpenAIEndpoint=$AZURE_OPENAI_ENDPOINT
+
 ARG AZURE_OPENAI_KEY
 ENV AzureOpenAIKey=$AZURE_OPENAI_KEY
 
@@ -10,10 +13,10 @@ ARG TBA_API_KEY
 ENV TBA_API_KEY=$TBA_API_KEY
 
 ARG OrchestratorEndpoint
-ENV OrchestratorEndpoint=$OrchestratorEndpoint
+ENV Yaap__Server__Endpoint=$OrchestratorEndpoint
 
 WORKDIR /app
 
 COPY . .
 
-ENTRYPOINT ["dotnet", "Teams_WS.dll"]
+ENTRYPOINT ["dotnet", "TeamsExpert.dll"]
