+++ samples/dotnet/websockets/Agents/TeamsExpert/TeamApi_Extras.cs
