+++ samples/dotnet/websockets/Agents/TeamsExpert/appsettings.Development.json
