+++ samples/dotnet/websockets/Agents/TeamsExpert/appsettings.json
