--- samples/dotnet/websockets/Agents/WebSockets/Teams_WS/Agent.cs
@@ -1,9 +1,0 @@
-﻿namespace Teams_WS;
-
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
-using Microsoft.SemanticKernel;
-
-using wsAgent.Core;
-
-internal class Agent(IConfiguration appConfig, ILoggerFactory loggerFactory, IHttpClientFactory httpClientFactory, Kernel sk, PromptExecutionSettings promptSettings) : Expert(appConfig, loggerFactory, httpClientFactory, sk, promptSettings) { }
