--- samples/dotnet/websockets/Agents/WebSockets/Teams_WS/Teams_WS.csproj
@@ -1,38 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-	<PropertyGroup>
-		<UserSecretsId>5631e549-948c-4903-be18-a06152c3600c</UserSecretsId>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Api\TeamApi.cs" Link="TeamApi.cs" />
-		<Compile Include="..\..\SignalR\Teams_SignalR\TeamApi_Extras.cs" Link="TeamApi_Extras.cs" />
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Model\Team.cs" Link="Models\Team.cs" />
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Model\TeamEventStatus.cs" Link="Models\TeamEventStatus.cs" />
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Model\TeamEventStatusAlliance.cs" Link="Models\TeamEventStatusAlliance.cs" />
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Model\TeamEventStatusAllianceBackup.cs" Link="Models\TeamEventStatusAllianceBackup.cs" />
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Model\TeamEventStatusPlayoff.cs" Link="Models\TeamEventStatusPlayoff.cs" />
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Model\TeamEventStatusRank.cs" Link="Models\TeamEventStatusRank.cs" />
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Model\TeamEventStatusRankRanking.cs" Link="Models\TeamEventStatusRankRanking.cs" />
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Model\TeamEventStatusRankSortOrderInfo.cs" Link="Models\TeamEventStatusRankSortOrderInfo.cs" />
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Model\TeamRobot.cs" Link="Models\TeamRobot.cs" />
-		<Compile Include="..\..\..\tba-client\src\TBAAPI.V3Client\Model\TeamSimple.cs" Link="Models\TeamSimple.cs" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<PackageReference Include="JsonCons.JmesPath" />
-		<PackageReference Include="Azure.Identity" />
-		<PackageReference Include="Microsoft.Extensions.Http" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\wsAgent.Core\wsAgent.Core.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<None Update="Dockerfile">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</None>
-	</ItemGroup>
-
-</Project>
