--- samples/dotnet/websockets/Agents/WebSockets/wsAgent.Core/wsAgent.Core.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<ItemGroup>
-		<PackageReference Include="Azure.Identity" />
-		<PackageReference Include="Microsoft.SemanticKernel" />
-		<PackageReference Include="System.Linq.Async" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\..\..\..\..\..\src\Client\Yaap.Client.csproj" />
-		<ProjectReference Include="..\..\..\ApiClient\ApiClient.csproj" />
-		<ProjectReference Include="..\..\..\Assistants\Assistants.csproj" />
-		<ProjectReference Include="..\..\..\Models\Models.csproj" />
-	</ItemGroup>
-
-</Project>
