+++ samples/dotnet/websockets/ConsoleApp/Log.cs
