+++ samples/dotnet/websockets/ConsoleApp/Program.cs
