+++ samples/dotnet/websockets/ConsoleApp/Properties/launchSettings.json
