+++ samples/dotnet/websockets/ConsoleApp/Worker.cs
