+++ samples/dotnet/websockets/ConsoleApp/appsettings.Development.json
