+++ samples/dotnet/websockets/ConsoleApp/appsettings.json
