+++ src/Client/Yaap.Client.Abstractions.csproj
@@ -10,7 +10,7 @@
     <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Models\Yaap.Core.csproj" />
+    <ProjectReference Include="..\Core\Yaap.Core.csproj" />
   </ItemGroup>
 
 </Project>
