+++ src/Core/Models/YaapClientDetail.cs
