+++ src/Core/Yaap.Core.csproj
