+++ src/Server/Yaap.Server.Abstractions.csproj
@@ -7,6 +7,6 @@
 		<PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
 	</ItemGroup>
 	<ItemGroup>
-	  <ProjectReference Include="..\Models\Yaap.Core.csproj" />
+	  <ProjectReference Include="..\Core\Yaap.Core.csproj" />
 	</ItemGroup>
 </Project>
